feat(models): Add locationSharing to User schema
This commit extends the `User` model to support a new user preference for location sharing.

A `locationSharing` boolean field has been added to the `preferences` object within the `UserSchema`. This field will default to `true`, allowing users to share their location by default.

This schema change is a prerequisite for the new location sharing preference feature, enabling the backend to store and manage this user setting.